Remove query and fragment parts of URLs from enclosure filenames.
Enclosures in RSS are linked to using standard URLs, which may have a
query part (starting with ?) or a fragment part (starting with #). We
need to trim this off before when determining the filename to use for
the enclosure. Ideally we would do proper URI parsing, but pulling in
another dependency for that seems like overkill so this patch instead
solves it using g_strsplit_set.

Fixes #4.